+++ lib/essentials/db/feature_level_providers/conversation_graph_connection_lifecycle_provider.g.dart
@@ -1,30 +1,30 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'message_data_reset_service.dart';
+part of 'conversation_graph_connection_lifecycle_provider.dart';
 
 // **************************************************************************
 // RiverpodGenerator
 // **************************************************************************
 
-String _$messageDataResetServiceHash() =>
-    r'1c0c85a7336bd19dc406cf858c99e799b1be3c4e';
+String _$conversationGraphConnectionLifecycleHash() =>
+    r'5c628bfc77d7c27f3f26eeebd4635840d7e1bd4d';
 
-/// See also [messageDataResetService].
-@ProviderFor(messageDataResetService)
-final messageDataResetServiceProvider =
-    AutoDisposeProvider<MessageDataResetService>.internal(
-      messageDataResetService,
-      name: r'messageDataResetServiceProvider',
+/// See also [conversationGraphConnectionLifecycle].
+@ProviderFor(conversationGraphConnectionLifecycle)
+final conversationGraphConnectionLifecycleProvider =
+    Provider<ConversationGraphConnectionLifecycle>.internal(
+      conversationGraphConnectionLifecycle,
+      name: r'conversationGraphConnectionLifecycleProvider',
       debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
           ? null
-          : _$messageDataResetServiceHash,
+          : _$conversationGraphConnectionLifecycleHash,
       dependencies: null,
       allTransitiveDependencies: null,
     );
 
 @Deprecated('Will be removed in 3.0. Use Ref instead')
 // ignore: unused_element
-typedef MessageDataResetServiceRef =
-    AutoDisposeProviderRef<MessageDataResetService>;
+typedef ConversationGraphConnectionLifecycleRef =
+    ProviderRef<ConversationGraphConnectionLifecycle>;
 // ignore_for_file: type=lint
 // ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package