+++ lib/features/settings/application/historical_archives_workflow_panel_model_provider.g.dart
@@ -27,7 +27,7 @@ final historicalArchivesWorkflowPanelModelProvider =
 typedef HistoricalArchivesWorkflowPanelModelRef =
     AutoDisposeProviderRef<HistoricalArchivesWorkflowPanelViewModel>;
 String _$historicalArchivesWorkflowHash() =>
-    r'4ecbd69aa3ed95a3bbc97053a54205cfda5a91fd';
+    r'e0cae5676886840a7bae1022bcc9b658c617cb69';
 
 /// See also [HistoricalArchivesWorkflow].
 @ProviderFor(HistoricalArchivesWorkflow)